Create the basic file that will read the file
This commit creates a basic simple QuakeLogParser::Parser class that
will be responsible to read the file specified as argument. It raises
an error if filename is not present.

bin/quake_log_parser was updated to call this class and now accepts a
filename as parameter like:
bin/quake_log_parser.rb qgames.txt