Create a Regex parse strategy
This commit creates a new parse strategy that will parse the log line
via regex and return a proper Command class with the additional
arguments found on the log line.

All implemented commands was created on lib/quake_log_parser/commands

If in future we want to go with another parsing strategy (like Treetop),
we can implement it and use on QuakeLogParser::Parser.